+++ Sourcery/main.swift
@@ -48,7 +48,7 @@ struct CustomArguments: ArgumentConvertible {
 fileprivate enum Validators {
     static func isReadable(path: Path) -> Path {
         if !path.isReadable {
-            print("'\(path)' does not exist or is not readable.")
+            Log.error("'\(path)' does not exist or is not readable.")
             exit(1)
         }
 
@@ -59,7 +59,7 @@ fileprivate enum Validators {
         _ = isReadable(path: path)
 
         if !(path.isDirectory || path.isFile) {
-            print("'\(path)' isn't a directory or proper file.")
+            Log.error("'\(path)' isn't a directory or proper file.")
             exit(2)
         }
 
@@ -68,7 +68,7 @@ fileprivate enum Validators {
 
     static func isWriteable(path: Path) -> Path {
         if path.exists && !path.isWritable {
-            print("'\(path)' isn't writeable.")
+            Log.error("'\(path)' isn't writeable.")
             exit(2)
         }
         return path
@@ -79,7 +79,7 @@ extension Configuration {
 
     func validate() {
         guard !source.isEmpty else {
-            print("No sources provided.")
+            Log.error("No sources provided.")
             exit(3)
         }
         if case let .sources(sources) = source {
@@ -87,7 +87,7 @@ extension Configuration {
         }
         _ = templates.map(Validators.isFileOrDirectory(path:))
         guard !templates.isEmpty else {
-            print("No templates provided.")
+            Log.error("No templates provided.")
             exit(3)
         }
         _ = Validators.isWriteable(path: output)
@@ -105,7 +105,10 @@ func runCLI() {
              description: "Stops using cache."),
         Flag("verbose",
              flag: "v",
-             description: "Turn on verbose logging for ignored entities"),
+             description: "Turn on verbose logging"),
+        Flag("quiet",
+             flag: "q",
+             description: "Turn off any logging, only emmit errors"),
         Flag("prune",
              flag: "p",
              description: "Remove empty generated files"),
@@ -114,13 +117,14 @@ func runCLI() {
         Option<Path>("output", ".", description: "Path to output. File or Directory. Default is current path."),
         Option<Path>("config", ".", description: "Path to config file. Directory. Default is current path."),
         Argument<CustomArguments>("args", description: "Custom values to pass to templates.")
-    ) { watcherEnabled, disableCache, verboseLogging, prune, sources, templates, output, configPath, args in
+    ) { watcherEnabled, disableCache, verboseLogging, quiet, prune, sources, templates, output, configPath, args in
         do {
+            Log.level = verboseLogging ? .verbose : quiet ? .errors : .warnings
             let configuration: Configuration
 
             let yamlPath: Path = configPath + ".sourcery.yml"
             if yamlPath.exists, let dict = try Yams.load(yaml: yamlPath.read()) as? [String: Any] {
-                configuration = Configuration(dict: dict)
+                configuration = Configuration(dict: dict, relativePath: configPath)
             } else {
                 configuration = Configuration(sources: sources,
                                               templates: templates,
@@ -140,10 +144,10 @@ func runCLI() {
                 RunLoop.current.run()
                 _ = keepAlive
             } else {
-                print("Processing time \(CFAbsoluteTimeGetCurrent() - start) seconds")
+                Log.info("Processing time \(CFAbsoluteTimeGetCurrent() - start) seconds")
             }
         } catch {
-            print(error)
+            Log.error("\(error)")
             exit(4)
         }
         }.run(Sourcery.version)