+++ SourceryTests/Parsing/ComposerSpec.swift
@@ -21,7 +21,7 @@ class ParserComposerSpec: QuickSpec {
             describe("uniqueType") {
                 func parse(_ code: String) -> [Type] {
                     guard let parserResult = try? FileParser(contents: code).parse() else { fail(); return [] }
-                    return Composer(verbose: false).uniqueTypes(parserResult)
+                    return Composer().uniqueTypes(parserResult)
                 }
 
                 context("given enum containing associated values") {