+++ SourceryTests/Parsing/FileParserSpec.swift
@@ -15,7 +15,7 @@ class FileParserSpec: QuickSpec {
             describe("parse") {
                 func parse(_ code: String) -> [Type] {
                     guard let parserResult = try? FileParser(contents: code).parse() else { fail(); return [] }
-                    return Composer(verbose: false).uniqueTypes(parserResult)
+                    return Composer().uniqueTypes(parserResult)
                 }
 
                 describe("regression files") {